cmd/snap-update-ns: move debug functions to dedicated file
